- move some Table code to a separate helper module #1785
